% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdns-package.R
\docType{package}
\name{getdns}
\alias{getdns}
\alias{getdns-package}
\title{Query Domain Name System Using the 'getdns' Application Programming Interface}
\description{
\href{https://getdnsapi.net/}{getdns} is a modern asynchronous DNS API. It
implements DNS entry points from a design developed and vetted by application
developers, in an API specification. The open source C implementation of
getdns is developed and maintained in collaboration by NLnet Labs, Sinodun
and No Mountain Software.
}
\author{
Bob Rudis (bob@rud.is)
}
